test(browser): migrate all product journeys to pinned Playwright
