+++ .github/workflows/ci.yml
@@ -352,7 +352,12 @@ jobs:
         with:
           node-version: "24"
           cache: npm
-          cache-dependency-path: lang/crates/vo-web/package-lock.json
+          cache-dependency-path: |
+            lang/crates/vo-web/package-lock.json
+            eng/browser/package-lock.json
+
+      - name: Setup browser contracts
+        uses: ./.github/actions/setup-browser
 
       - name: Cache wasm-pack
         uses: actions/cache@55cc8345863c7cc4c66a329aec7e433d2d1c52a9 # v6.1.0
@@ -592,6 +597,15 @@ jobs:
           name: ci-evidence-wasm-web-${{ github.sha }}
           path: target/ci/evidence/*.evidence.json
           if-no-files-found: error
+          retention-days: 14
+
+      - name: Upload browser diagnostics
+        if: always()
+        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+        with:
+          name: browser-diagnostics-${{ github.sha }}
+          path: target/ci/browser
+          if-no-files-found: ignore
           retention-days: 14
 
       - name: Upload promotable Studio site
