+++ docs/ci.md
@@ -169,6 +169,14 @@ npm --prefix lang/crates/vo-web run test:ui
 npm --prefix lang/crates/vo-web run test:ui-browser
 ```
 
+Browser scenarios use the exact Playwright version and Chromium revision in
+`eng/browser/package-lock.json`. The compatibility entry point preserves all
+eight product scenarios, with their assertion mapping in
+`eng/browser/coverage.json`. Each attempt retains failure traces, screenshots,
+console/network diagnostics, and HTML/JSON reports under `target/ci/browser`.
+Retries are disabled; `eng/browser/check-diagnostics.mjs` verifies controlled
+failures for all eight scenarios. See `eng/browser/README.md` for local setup.
+
 Dependency audits are:
 
 ```sh
