+++ eng/browser/coverage.json
@@ -1,0 +1,243 @@
+{
+  "schema": "volang.browser-contract-migration.v1",
+  "baseline": "552556b5",
+  "scenarios": [
+    {
+      "id": "ui-conformance",
+      "source_function": "runSmoke",
+      "contract": "lang/crates/vo-web/js/browser_smoke.ts",
+      "checks": [
+        "primitive and visual style mapping",
+        "form accessibility mapping",
+        "typed browser input event encoding",
+        "IME composition ownership",
+        "controlled UTF-16 synchronization",
+        "modal focus and Stack projection",
+        "modal forward focus wrapping",
+        "modal background pointer isolation",
+        "modal focus restoration",
+        "atomic rejection of ambiguous modal trees"
+      ],
+      "flag": "ui-conformance"
+    },
+    {
+      "id": "data-application",
+      "source_function": "runDataApplicationSmoke",
+      "source_lines": 243,
+      "contract": "eng/browser/scenarios/data-application.mjs",
+      "checkpoints": [
+        "initial",
+        "selection",
+        "optimistic",
+        "committed",
+        "page",
+        "filter",
+        "sort",
+        "offline",
+        "commands",
+        "commandActivated",
+        "settings",
+        "validation",
+        "submission",
+        "restored"
+      ],
+      "flag": "dataApplicationSmoke"
+    },
+    {
+      "id": "content-site",
+      "source_function": "runContentSiteSmoke",
+      "source_lines": 277,
+      "contract": "eng/browser/scenarios/content-site.mjs",
+      "checkpoints": [
+        "initial",
+        "article",
+        "search",
+        "validation",
+        "invalidEmail",
+        "validEmail",
+        "validFields",
+        "submission",
+        "offline",
+        "productionFetch"
+      ],
+      "flag": "contentSiteSmoke"
+    },
+    {
+      "id": "media-application",
+      "source_function": "runMediaApplicationSmoke",
+      "source_lines": 178,
+      "contract": "eng/browser/scenarios/media-application.mjs",
+      "checkpoints": [
+        "initial",
+        "play",
+        "seek",
+        "volume",
+        "rate",
+        "visualizations",
+        "permission",
+        "capturing",
+        "stopped",
+        "denied",
+        "recovered",
+        "offline",
+        "productionFetch"
+      ],
+      "flag": "mediaApplicationSmoke"
+    },
+    {
+      "id": "uikit-gallery",
+      "source_function": "runUikitGallerySmoke",
+      "source_lines": 1363,
+      "contract": "eng/browser/scenarios/uikit-gallery.mjs",
+      "checkpoints": [
+        "initial",
+        "interactionVisuals",
+        "search",
+        "searchCleared",
+        "numberInput",
+        "segmented",
+        "list",
+        "tag",
+        "inputGroup",
+        "codeSnippet",
+        "comboboxFiltered",
+        "comboboxSelected",
+        "multiSelectFiltered",
+        "multiSelectSelected",
+        "multiSelectRemoved",
+        "datePickerParsed",
+        "calendarRangeRestarted",
+        "calendarKeyboard",
+        "uploadCancelled",
+        "uploadRetried",
+        "uploadDropped",
+        "uploadTypeRejected",
+        "contextMenuOpened",
+        "contextMenuTypeahead",
+        "contextSubmenu",
+        "contextMenuActivated",
+        "contextMenuDismissed",
+        "contextMenuHover",
+        "contextMenuPointerActivated",
+        "navigationOverflow",
+        "navigationCompact",
+        "navigationSide",
+        "toggleRovingFocus",
+        "toggleGroup",
+        "rating",
+        "resizableKeyboard",
+        "resizablePointer",
+        "resizableCollapsed",
+        "carousel",
+        "carouselPointer",
+        "scrollArea",
+        "checkbox",
+        "textArea",
+        "slider",
+        "accordion",
+        "menu",
+        "menuToggle",
+        "dataSort",
+        "dataSelection",
+        "treeCollapsed",
+        "treeExpanded",
+        "toastVisible",
+        "toastDismissed",
+        "drawer",
+        "environment",
+        "environmentRestored",
+        "responsive",
+        "commandFiltered",
+        "command"
+      ],
+      "flag": "uikitGallerySmoke",
+      "added_checkpoints": [
+        "commandPointer",
+        "commandRowEscape",
+        "commandInputEscape"
+      ],
+      "regressions": "Pointer activation of a non-initial command and Escape from input/row; Enter closes without reactivating restored focus."
+    },
+    {
+      "id": "component-state",
+      "source_function": "runComponentStateSmoke",
+      "source_lines": 67,
+      "contract": "eng/browser/scenarios/component-state.mjs",
+      "checkpoints": [
+        "initial",
+        "afterCounterClick",
+        "afterMovement",
+        "afterRemoval",
+        "afterInsertion",
+        "afterReplacement"
+      ],
+      "flag": "componentStateSmoke"
+    },
+    {
+      "id": "studio-workbench",
+      "source_function": "runStudioWorkbenchSmoke",
+      "source_lines": 273,
+      "contract": "eng/browser/scenarios/studio-workbench.mjs",
+      "checkpoints": [
+        "initial",
+        "edited",
+        "saved",
+        "fileSwitch",
+        "run",
+        "console",
+        "commandFiltered",
+        "palette",
+        "documentation",
+        "restored",
+        "offline",
+        "productionFetch"
+      ],
+      "flag": "studioWorkbenchSmoke"
+    },
+    {
+      "id": "studio",
+      "source_function": "runStudioAotSmoke",
+      "source_lines": 1014,
+      "contract": "eng/browser/scenarios/studio.mjs",
+      "checkpoints": [
+        "home",
+        "performance",
+        "compactLayout",
+        "projectValidation",
+        "edited",
+        "editorScroll",
+        "unsavedSearch",
+        "noOverwriteCreate",
+        "nestedSourcePackage",
+        "directoryCollapse",
+        "deleteConfirmation",
+        "run",
+        "runCancellation",
+        "commandOverlay",
+        "command",
+        "documentation",
+        "languageExample",
+        "languageExampleOpenMs",
+        "uiExampleOpenMs",
+        "diagnosticNavigation",
+        "dirtyRevert",
+        "cleanSaveDisabled",
+        "preview",
+        "interaction",
+        "previewMessageIsolation",
+        "account",
+        "accountCredential",
+        "accountCancellation",
+        "share",
+        "runner",
+        "warmReloadMs",
+        "runnerColdStart",
+        "runnerReturn",
+        "sharedProjectPersistence",
+        "sharedProjectReopen",
+        "documentationColdStart"
+      ],
+      "flag": "studioAotSmoke"
+    }
+  ]
+}
